nordismvault/requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.7
django-waffle==2.1.0
djangorestframework==3.12.2
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
iniconfig==1.1.1
oauthlib==3.1.0
packaging==20.9
pluggy==0.13.1
protobuf==3.15.6
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==6.2.2
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
pytest-django==4.1.0
